Introduce code review feature in GithubEventHandler
The GithubEventHandler class now includes a Reviewer instance, responsible for reviewing file changes. For every pull request, the class will fetch the latest commit SHA, retrieve the files changed, and initiate a review of these changes. Any comments generated as part of the review are then posted as comments on the appropriate lines of the mentioned files.